requirements.txt
fastapi
uvicorn
python-dotenv
dashscope